+++ src/main/java/com/anyang/maruni/domain/notification/infrastructure/decorator/NotificationHistoryDecorator.java
@@ -1,4 +1,4 @@
-package com.anyang.maruni.domain.notification.infrastructure;
+package com.anyang.maruni.domain.notification.infrastructure.decorator;
 
 import com.anyang.maruni.domain.notification.domain.entity.NotificationHistory;
 import com.anyang.maruni.domain.notification.domain.service.NotificationHistoryService;
