+++ src/test/java/com/anyang/maruni/domain/notification/infrastructure/FirebasePushNotificationServiceTest.java
@@ -3,9 +3,10 @@ package com.anyang.maruni.domain.notification.infrastructure;
 import com.anyang.maruni.domain.notification.domain.exception.NotificationException;
 import com.anyang.maruni.domain.notification.domain.vo.NotificationChannelType;
 import com.anyang.maruni.domain.notification.domain.service.PushTokenService;
+import com.anyang.maruni.domain.notification.infrastructure.firebase.FirebaseMessagingWrapper;
+import com.anyang.maruni.domain.notification.infrastructure.service.FirebasePushNotificationService;
 import com.anyang.maruni.global.config.properties.FirebaseProperties;
 import com.anyang.maruni.global.response.error.ErrorCode;
-import com.google.firebase.messaging.FirebaseMessagingException;
 import com.google.firebase.messaging.Message;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
